% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{osm_classify}
\alias{osm_classify}
\title{Classify OSM Features}
\usage{
osm_classify(data, classification)
}
\arguments{
\item{data}{imported layer from an OSM PBF file.}

\item{classification}{a 2-level nested list providing a classification. The layers of the list are:
\tabular{ll}{
  categories \tab a list of tags and values that make up a feature category \cr
  tags \tab a character vector of tag values to match on, or \code{""} to match all possible tags
}}
}
\description{
Classify OSM Features
}
\examples{
str(osm_line_class)
str(osm_point_polygon_class)
}
