% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifications.R
\docType{data}
\name{osm_line_class}
\alias{osm_line_class}
\title{Functional Classification of OSM Line Features}
\format{
An object of class \code{list} of length 11.
}
\usage{
osm_line_class
}
\description{
A nested list proposing a classification of OSM lines into 11 functional categories.
}
\examples{
collapse::unlist2d(osm_line_class, idcols = c("category", "tag"))
}
\seealso{
\link{osmclass-package}
}
\keyword{datasets}
