% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifications.R
\docType{data}
\name{osm_line_info_tags}
\alias{osm_line_info_tags}
\title{Further tags of Interest for Line Features}
\format{
An object of class \code{list} of length 11.
}
\usage{
osm_line_info_tags
}
\description{
Further tags that max be of interested to extract for OSM line features.
}
\examples{
collapse::unlist2d(osm_line_info_tags, idcols = "category")
}
\seealso{
\link{osmclass-package}
}
\keyword{datasets}
