% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{osm_other_tags_list}
\alias{osm_other_tags_list}
\title{Generate a List from the 'other_tags' Column in OSM PBF Data}
\usage{
osm_other_tags_list(x, values = FALSE, split = "\\",\\"|\\"=>\\"", ...)
}
\arguments{
\item{x}{character. The 'other_tags' column of an imported osm.pbf file.}

\item{values}{logical. \code{TRUE} also includes the values of tags.}

\item{split}{character. Pattern passed to \code{\link{strsplit}} to split up \code{x}.}

\item{\dots}{further arguments to \code{\link{strsplit}}.}
}
\value{
a list of tags as character vectors, or a nested list of tags and values if \code{values = TRUE}.
}
\description{
Generate a List from the 'other_tags' Column in OSM PBF Data
}
\examples{
# See Examples at ?osmclass

}
\seealso{
\link{osmclass-package}
}
