% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{package}
\name{osmclass-package}
\alias{osmclass-package}
\alias{osmclass}
\title{Classify Open Street Map Features}
\description{
An R package to classify Open Street Map (OSM) features into meaningful functional or analytical categories.
It expects OSM PBF data, e.g. from https://download.geofabrik.de/, imported as \emph{sf} data frames, and
is well optimized to deal with large quantities of such data. \cr\cr
}
\section{Functions}{

Main Function to Classify OSM Features

\code{\link[=osm_classify]{osm_classify()}}

Auxiliary Functions to Extract Information (Tags) from OSM PBF Layers

\code{\link[=osm_other_tags_list]{osm_other_tags_list()}}\cr
\code{\link[=osm_tags_df]{osm_tags_df()}}\cr\cr
}

\section{Classifications}{

Suggested OSM Classsification, developed for the Africa OSM Following Krantz (2023)

\code{\link{osm_point_polygon_class}}\cr
\code{\link{osm_line_class}}\cr
\code{\link{osm_line_info_tags}}
}

