% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{osm_classify}
\alias{osm_classify}
\title{Classify OSM Features}
\usage{
osm_classify(data, classification)
}
\arguments{
\item{data}{imported layer from an OSM PBF file.}

\item{classification}{a 2-level nested list providing a classification. The layers of the list are:
\tabular{ll}{
  categories \tab a list of tags and values that make up a feature category \cr\cr\cr
  tags \tab a character vector of tag values to match on, or \code{""} to match all possible tag values.
  It is also possible to match all except certain tags by negating them with \code{"!"} e.g. \code{"!no"}. \cr
}}
}
\description{
Classify OSM Features
}
\note{
It is not necessary to expand the 'other_tags' column, e.g. using \code{\link{osm_tags_df}}. \code{osm_classify()} efficiently searches the content of that column without expanding it.
}
\examples{
str(osm_line_class)
str(osm_point_polygon_class)
}
