% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifications.R
\docType{data}
\name{osm_point_polygon_class}
\alias{osm_point_polygon_class}
\title{Functional Classification of OSM Points and Polygon Features}
\format{
An object of class \code{list} of length 33.
}
\usage{
osm_point_polygon_class
}
\description{
A nested list proposing a classification of OSM point and polygons into 33 functional categories.
It was developed to accurately classify OSM features in Africa, see Krantz (2023).
}
\examples{
collapse::unlist2d(osm_point_polygon_class, idcols = c("category", "tag"))
}
\keyword{datasets}
