% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{osm_tags_df}
\alias{osm_tags_df}
\title{Extract Tags and Values as Columns from an OSM PBF Data}
\usage{
osm_tags_df(data, tags, na.prop = 0)
}
\arguments{
\item{data}{an imported layer from an OSM PBF file.}

\item{tags}{character. A vector of tags to extract as columns.}

\item{na.prop}{double. Proportion of features having a tag in order to keep the column.}
}
\description{
Extract Tags and Values as Columns from an OSM PBF Data
}
